[SI-41] Chore: GameSetting.java change directory
